Pushing everything to run on the checkers
