+++ bytecodes/06.m
@@ -1,0 +1,6 @@
+push 1
+pint
+push 2
+pint
+push 3
+pint
