+++ monty.h
@@ -38,6 +38,8 @@ typedef struct instruction_s
 /* Prototypes Section */
 int main(int ac, char **av);
 char **tokenization(char *ptr, char *delim);
+
+/* Operation cmd */
 void push(stack_t **stack, unsigned int line_number);
 void pall(stack_t **stack, unsigned int line_number);
 void pint(stack_t **stack, unsigned int line_number);
@@ -45,7 +47,11 @@ void pop(stack_t **stack, unsigned int line_number);
 void nop(stack_t **stack, unsigned int line_number);
 void swap(stack_t **stack, unsigned int line_number);
 void add(stack_t **stack, unsigned int line_number);
+
+/* Function that free */
 void free_array(char **tokens);
 void free_stack(stack_t *stack);
+
+/* Function to combine our cmd with the function */
 void (*get_op_func(char *cmd))(stack_t **stack, unsigned int line_number);
 #endif
